fix vai loz
